feat: document PlayerAvatarExtendedInfoWriter
