+++ protocol/osrs-221-model/src/main/kotlin/net/rsprot/protocol/game/outgoing/info/playerinfo/PlayerInfo.kt
@@ -55,7 +55,7 @@ public class PlayerInfo internal constructor(
     private val allocator: ByteBufAllocator,
     private var platformType: PlatformType,
     extendedInfoFilter: ExtendedInfoFilter,
-    extendedInfoWriters: List<AvatarExtendedInfoWriter>,
+    extendedInfoWriters: List<PlayerAvatarExtendedInfoWriter>,
     huffmanCodec: HuffmanCodec,
 ) : ReferencePooledObject, OutgoingMessage {
     /**
